rf: `translateArrayTypeAnnotation` fns > common fn
- combined `Flow` and `TS` `translateArrayTypeAnnotation` fn 's into common fn
- moved it to `parsers-primitives.js`
- re-organized imports and exports

Extras:
- resolved conflicts :)

Temp:

[skip ci]
